src/main/java/com/othkkartho/vtuber_site_v2/service/membermanage/MemberService.java
package com.othkkartho.vtuber_site_v2.service.membermanage;

import com.othkkartho.vtuber_site_v2.dto.membermanage.member.MemberDto;
import com.othkkartho.vtuber_site_v2.exception.MemberNotFoundException;
import com.othkkartho.vtuber_site_v2.repository.member.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class MemberService {
    private final MemberRepository memberRepository;

    public MemberDto read(Long id) {
        return MemberDto.toEntity(memberRepository.findById(id).orElseThrow(MemberNotFoundException::new));
    }

    @Transactional
    public void delete(Long id) {
        if(notExistsMember(id)) throw new MemberNotFoundException();
        memberRepository.deleteById(id);
    }

    private boolean notExistsMember(Long id) {
        return !memberRepository.existsById(id);
    }
}
